Move policy to separate module
